autoscaling/app/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
psycopg2-binary>=2.9.9
databricks-sdk>=0.96.0
pydantic>=2.0.0
